+++ topology/point_3d.hpp
@@ -362,7 +362,7 @@ namespace scifir
 
 			float get_p() const
 			{
-				return std::sqrt(std::pow(x,2) + std::pow(y,2));
+				return float(std::sqrt(std::pow(x,2) + std::pow(y,2)));
 			}
 
 			angle get_theta() const
@@ -372,7 +372,7 @@ namespace scifir
 
 			float get_r() const
 			{
-				return std::sqrt(std::pow(x,2) + std::pow(y,2) + std::pow(z,2));
+				return float(std::sqrt(std::pow(x,2) + std::pow(y,2) + std::pow(z,2)));
 			}
 
 			angle get_phi() const
@@ -477,7 +477,7 @@ namespace scifir
 
 			float distance_to_origin() const
 			{
-				return std::sqrt(std::pow(x,2) + std::pow(y,2) + std::pow(z,2));
+				return float(std::sqrt(std::pow(x,2) + std::pow(y,2) + std::pow(z,2)));
 			}
 
 			string display_cylindrical() const
